Move AlignmentPicker to a separate component.
