+++ experiments/tester/src/RNTester/TestComponents/PersonaCoin/styles.ts
@@ -14,7 +14,7 @@ export const styles = StyleSheet.create({
     alignItems: 'center'
   },
   header: {
-    marginTop: 12,
+    marginVertical: 6,
     fontSize: 12
   },
   section: {
